requirements.txt
asgiref                 3.5.2
async-timeout           4.0.2
certifi                 2022.9.24
cffi                    1.15.1
charset-normalizer      2.1.1
coreapi                 2.3.3
coreschema              0.0.4
cryptography            38.0.1
defusedxml              0.7.1
Deprecated              1.2.13
Django                  4.1.1
django-allauth          0.51.0
django-cors-headers     3.13.0
django-crispy-forms     1.14.0
django-filter           22.1
django-formtools        2.4
django-guardian         2.4.0
django-redis            5.2.0
django-reversion        5.0.3
django-utils-six        2.0
djangorestframework     3.14.0
djangorestframework-jwt 1.11.0
drf-extensions          0.7.1
Faker                   15.0.0
future                  0.18.2
httplib2                0.20.4
idna                    3.4
itypes                  1.2.0
Jinja2                  3.1.2
Markdown                3.4.1
MarkupSafe              2.1.1
mysqlclient             2.1.1
oauthlib                3.2.1
packaging               21.3
Pillow                  9.2.0
pip                     22.2.2
pycparser               2.21
pycryptodome            3.15.0
PyJWT                   1.7.1
pyparsing               3.0.9
python-dateutil         2.8.2
python3-openid          3.2.0
pytz                    2022.4
raven                   6.10.0
redis                   4.3.4
requests                2.28.1
requests-oauthlib       1.3.1
setuptools              65.3.0
six                     1.16.0
sqlparse                0.4.3
uritemplate             4.1.1
urllib3                 1.26.12
wheel                   0.37.1
wrapt                   1.14.1
XlsxWriter              3.0.3
xlwt                    1.3.0
